Refactor stop area form input mapping
Pass around form state instead of converting it to input in form.
This is similar to how we do with stop place forms.
Makes it easier to handle Tiamat types when we only need to deal with
the conversion in the hook right before sending the request.
